fix(board): one active indicator in the lens picker, aligned
The saved-view checkmark could show alongside the base lens's checkmark
(e.g. an "All + filter" saved view lit up both All and the view) and sat
indented from the lens-list checks because the hover actions reserved
space to its right.

- Resolve the active saved view once in BoardLensMenu and suppress the
  lens check when one matches (a saved view IS the selection, not its
  base lens) — exactly one row reads as active.
- Align the saved-view check to the row's right edge (same column as the
  lens check) and cross-fade it with the rename/delete actions on
  hover/focus, so the right slot holds one thing at a time. Actions stay
  keyboard-reachable via focus-visible/focus-within.
